extract storage and origins methods from sessions manager
