+++ packages/driver/src/cy/commands/sessions/manager.ts
@@ -1,25 +1,11 @@
-import _ from 'lodash'
-import { $Location } from '../../../cypress/location'
 import type { ServerSessionData } from '@packages/types'
-import {
-  getCurrentOriginStorage,
-  setPostMessageLocalStorage,
-  getPostMessageLocalStorage,
-} from './utils'
+import _ from 'lodash'
+
+import { getAllHtmlOrigins } from './origins'
+import { clearStorage, getStorage, setStorage } from './storage'
 
 type ActiveSessions = Cypress.Commands.Session.ActiveSessions
 type SessionData = Cypress.Commands.Session.SessionData
-
-interface OriginStorageOptions {
-  clear?: boolean
-  origin: string
-  value?: any
-}
-
-interface SetStorageOptions {
-  localStorage?: OriginStorageOptions[]
-  sessionStorage?: OriginStorageOptions[]
-}
 
 const LOGS = {
   clearCurrentSessionData: {
@@ -74,66 +60,6 @@ export default class SessionsManager {
     this.cy.state('activeSessions', clearedSessions)
   }
 
-  mapOrigins = async (origins: string | Array<string>): Promise<Array<string>> => {
-    const getOrigins = this.Cypress.Promise.map(
-      ([] as string[]).concat(origins), async (v) => {
-        if (v === '*') {
-          return await this.getAllHtmlOrigins()
-        }
-
-        if (v === 'currentOrigin') {
-          return $Location.create(window.location.href).origin
-        }
-
-        return $Location.create(v).origin
-      },
-    )
-
-    return _.uniq(_.flatten(await getOrigins))
-  }
-
-  _setStorageOnOrigins = async (originOptions) => {
-    const specWindow = this.cy.state('specWindow')
-
-    const currentOrigin = $Location.create(window.location.href).origin
-
-    const currentOriginIndex = _.findIndex(originOptions, { origin: currentOrigin })
-
-    if (currentOriginIndex !== -1) {
-      const opts = originOptions.splice(currentOriginIndex, 1)[0]
-
-      if (!_.isEmpty(opts.localStorage)) {
-        if (opts.localStorage.clear) {
-          window.localStorage.clear()
-        }
-
-        _.each(opts.localStorage.value, (val, key) => localStorage.setItem(key, val))
-      }
-
-      if (opts.sessionStorage) {
-        if (opts.sessionStorage.clear) {
-          window.sessionStorage.clear()
-        }
-
-        _.each(opts.sessionStorage.value, (val, key) => sessionStorage.setItem(key, val))
-      }
-    }
-
-    if (_.isEmpty(originOptions)) {
-      return
-    }
-
-    await setPostMessageLocalStorage(specWindow, originOptions)
-  }
-
-  getAllHtmlOrigins = async () => {
-    const currentOrigin = $Location.create(window.location.href).origin
-    const storedOrigins = await this.Cypress.backend('get:rendered:html:origins')
-    const origins = [..._.keys(storedOrigins), currentOrigin]
-
-    return _.uniq(origins)
-  }
-
   // this the public api exposed to consumers as Cypress.session
   sessions = {
     defineSession: (options = {} as any): SessionData => {
@@ -167,7 +93,7 @@ export default class SessionsManager {
       window.sessionStorage.clear()
 
       await Promise.all([
-        this.sessions.clearStorage(),
+        clearStorage(this.Cypress),
         this.sessions.clearCookies(),
       ])
     },
@@ -181,7 +107,7 @@ export default class SessionsManager {
     },
 
     setSessionData: async (data) => {
-      const allHtmlOrigins = await this.getAllHtmlOrigins()
+      const allHtmlOrigins = await getAllHtmlOrigins(this.Cypress)
 
       let _localStorage = data.localStorage || []
       let _sessionStorage = data.sessionStorage || []
@@ -197,7 +123,7 @@ export default class SessionsManager {
       })
 
       await Promise.all([
-        this.sessions.setStorage({ localStorage: _localStorage, sessionStorage: _sessionStorage }),
+        setStorage(this.Cypress, { localStorage: _localStorage, sessionStorage: _sessionStorage }),
         this.sessions.setCookies(data.cookies),
       ])
     },
@@ -216,7 +142,7 @@ export default class SessionsManager {
 
     getCurrentSessionData: async () => {
       const [storage, cookies] = await Promise.all([
-        this.sessions.getStorage({ origin: '*' }),
+        getStorage(this.Cypress, { origin: '*' }),
         this.sessions.getCookies(),
       ])
 
@@ -228,104 +154,6 @@ export default class SessionsManager {
 
     getSession: (id: string): Promise<ServerSessionData> => {
       return this.Cypress.backend('get:session', id)
-    },
-
-    /**
-      * 1) if we only need currentOrigin localStorage, access sync
-      * 2) if cross-origin http, we need to load in iframe from our proxy that will intercept all http reqs at /__cypress/automation/*
-      *      and postMessage() the localStorage value to us
-      * 3) if cross-origin https, since we pass-thru https connections in the proxy, we need to
-      *      send a message telling our proxy server to intercept the next req to the https domain,
-      *      then follow 2)
-      */
-    getStorage: async (options: InternalCypress.StorageOptions = {}): Promise<Cypress.Storages> => {
-      const specWindow = this.cy.state('specWindow')
-
-      if (!_.isObject(options)) {
-        throw new Error('getStorage() takes an object')
-      }
-
-      const opts = _.defaults({}, options, {
-        origin: 'currentOrigin',
-      })
-
-      const currentOrigin = $Location.create(window.location.href).origin
-
-      const origins: Array<string> = await this.mapOrigins(opts.origin)
-
-      const results = {
-        localStorage: [] as Cypress.OriginStorage[],
-        sessionStorage: [] as Cypress.OriginStorage[],
-      }
-
-      function pushValue (origin, value) {
-        if (!_.isEmpty(value.localStorage)) {
-          results.localStorage.push({ origin, value: value.localStorage })
-        }
-
-        if (!_.isEmpty(value.sessionStorage)) {
-          results.sessionStorage.push({ origin, value: value.sessionStorage })
-        }
-      }
-
-      const currentOriginIndex = origins.indexOf(currentOrigin)
-
-      if (currentOriginIndex !== -1) {
-        origins.splice(currentOriginIndex, 1)
-        const currentOriginStorage = getCurrentOriginStorage()
-
-        pushValue(currentOrigin, currentOriginStorage)
-      }
-
-      if (_.isEmpty(origins)) {
-        return results
-      }
-
-      if (currentOrigin.startsWith('https:')) {
-        _.remove(origins, (v) => v.startsWith('http:'))
-      }
-
-      const postMessageResults = await getPostMessageLocalStorage(specWindow, origins)
-
-      postMessageResults.forEach((val) => {
-        pushValue(val[0], val[1])
-      })
-
-      return results
-    },
-
-    clearStorage: async (type?: InternalCypress.StorageType) => {
-      const origins = await this.getAllHtmlOrigins()
-      const originOptions = origins.map((origin) => ({ origin, clear: true }))
-      const options: SetStorageOptions = {}
-
-      if (!type || type === 'localStorage') {
-        options.localStorage = originOptions
-      }
-
-      if (!type || type === 'sessionStorage') {
-        options.sessionStorage = originOptions
-      }
-
-      await this.sessions.setStorage(options)
-    },
-
-    setStorage: async (options: SetStorageOptions) => {
-      const currentOrigin = $Location.create(window.location.href).origin as string
-
-      const mapToCurrentOrigin = (v) => ({ ...v, origin: (v.origin && v.origin !== 'currentOrigin') ? $Location.create(v.origin).origin : currentOrigin })
-
-      const mappedLocalStorage = _.map(options.localStorage, (v) => {
-        return mapToCurrentOrigin({ origin: v.origin, localStorage: _.pick(v, 'value', 'clear') })
-      })
-
-      const mappedSessionStorage = _.map(options.sessionStorage, (v) => {
-        return mapToCurrentOrigin({ origin: v.origin, sessionStorage: _.pick(v, 'value', 'clear') })
-      })
-
-      const storageOptions = _.map(_.groupBy(mappedLocalStorage.concat(mappedSessionStorage), 'origin'), (v) => _.merge({}, ...v))
-
-      await this._setStorageOnOrigins(storageOptions)
     },
   }
 }
